requirements.txt
python-dotenv
google.generativeai
openai
google
pyodbc
chromadb  <-- Bunu ekleyin